# Further performance improvements