--- a/src/BlazorDatasheet.Core/Data/Cells/CellStoreRestoreData.cs
+++ b/src/BlazorDatasheet.Core/Data/Cells/CellStoreRestoreData.cs
@@ -11,13 +11,96 @@ namespace BlazorDatasheet.Core.Data.Cells;
 
 internal class CellStoreRestoreData
 {
-    internal MatrixRestoreData<CellValue> ValueRestoreData { get; set; } = new();
-    internal MatrixRestoreData<CellFormula?> FormulaRestoreData { get; set; } = new();
-    internal MatrixRestoreData<bool?> ValidRestoreData { get; set; } = new();
-    internal RegionRestoreData<string> TypeRestoreData { get; set; } = new();
-    internal RegionRestoreData<CellFormat> FormatRestoreData { get; set; } = new();
-    internal RegionRestoreData<bool> MergeRestoreData { get; set; } = new();
-    internal DependencyManagerRestoreData DependencyManagerRestoreData { get; set; } = new();
+    private MatrixRestoreData<CellValue>? _valueRestoreData;
+
+    internal MatrixRestoreData<CellValue> ValueRestoreData
+    {
+        get
+        {
+            if (_valueRestoreData == null)
+                _valueRestoreData = new();
+            return _valueRestoreData;
+        }
+        set => _valueRestoreData = value;
+    }
+
+    private MatrixRestoreData<CellFormula?>? _formulaRestoreData;
+
+    internal MatrixRestoreData<CellFormula?> FormulaRestoreData
+    {
+        get
+        {
+            if (_formulaRestoreData == null)
+                _formulaRestoreData = new();
+            return _formulaRestoreData;
+        }
+        set => _formulaRestoreData = value;
+    }
+
+    private MatrixRestoreData<bool?>? _validRestoreData;
+
+    internal MatrixRestoreData<bool?> ValidRestoreData
+    {
+        get
+        {
+            if (_validRestoreData == null)
+                _validRestoreData = new();
+            return _validRestoreData;
+        }
+        set => _validRestoreData = value;
+    }
+
+    private RegionRestoreData<string>? _typeRestoreData;
+
+    internal RegionRestoreData<string> TypeRestoreData
+    {
+        get
+        {
+            if (_typeRestoreData == null)
+                _typeRestoreData = new();
+            return _typeRestoreData;
+        }
+        set => _typeRestoreData = value;
+    }
+
+    private RegionRestoreData<CellFormat>? _formatRestoreData;
+
+    internal RegionRestoreData<CellFormat> FormatRestoreData
+    {
+        get
+        {
+            if (_formatRestoreData == null)
+                _formatRestoreData = new();
+            return _formatRestoreData;
+        }
+        set => _formatRestoreData = value;
+    }
+
+    private RegionRestoreData<bool>? _mergeRestoreData;
+
+    internal RegionRestoreData<bool> MergeRestoreData
+    {
+        get
+        {
+            if (_mergeRestoreData == null)
+                _mergeRestoreData = new();
+            return _mergeRestoreData;
+        }
+        set => _mergeRestoreData = value;
+    }
+
+    private DependencyManagerRestoreData? _dependencyGraph;
+
+    internal DependencyManagerRestoreData DependencyManagerRestoreData
+    {
+        get
+        {
+            if (_dependencyGraph == null)
+                _dependencyGraph = new();
+            return _dependencyGraph;
+        }
+        set => _dependencyGraph = value;
+    }
 
     internal IEnumerable<CellPosition> GetAffectedPositions()
     {
@@ -36,12 +119,25 @@ internal class CellStoreRestoreData
 
     internal void Merge(CellStoreRestoreData item)
     {
-        ValueRestoreData.Merge(item.ValueRestoreData);
-        FormulaRestoreData.Merge(item.FormulaRestoreData);
-        ValidRestoreData.Merge(item.ValidRestoreData);
-        TypeRestoreData.Merge(item.TypeRestoreData);
-        FormatRestoreData.Merge(item.FormatRestoreData);
-        MergeRestoreData.Merge(item.MergeRestoreData);
-        DependencyManagerRestoreData.Merge(item.DependencyManagerRestoreData);
+        if (item._valueRestoreData != null)
+            ValueRestoreData.Merge(item._valueRestoreData);
+
+        if (item._formulaRestoreData != null)
+            FormulaRestoreData.Merge(item._formulaRestoreData);
+
+        if (item._validRestoreData != null)
+            ValidRestoreData.Merge(item.ValidRestoreData);
+
+        if (item._typeRestoreData != null)
+            TypeRestoreData.Merge(item._typeRestoreData);
+
+        if (item._formatRestoreData != null)
+            FormatRestoreData.Merge(item._formatRestoreData);
+
+        if (item._mergeRestoreData != null)
+            MergeRestoreData.Merge(item.MergeRestoreData);
+
+        if (item._dependencyGraph != null)
+            DependencyManagerRestoreData.Merge(item._dependencyGraph);
     }
 }