… src/BlazorDatasheet.DataStructures/RTree/RTree.Utilities.cs
… src/BlazorDatasheet.DataStructures/RTree/RTree.cs
… src/BlazorDatasheet.DataStructures/Store/Range1DStore.cs
… src/BlazorDatasheet.DataStructures/Store/RegionDataStore.cs
… src/BlazorDatasheet.DataStructures/Util/SheetMath.cs
… src/BlazorDatasheet.Formula.Core/Dependencies/DependencyManager.cs
… test/BlazorDatasheet.Test/Geometry/IntervalsTest.cs
… test/BlazorDatasheet.Test/Store/RegionStoreTests.cs