Baseline
| Method         | Mean     | Error    | StdDev  | Allocated |
|--------------- |---------:|---------:|--------:|----------:|
| Set200x200Formula | 574.7 ms | 10.72 ms | 9.50 ms | 657.94 MB |

After the changes
| Method         | Mean     | Error   | StdDev   | Allocated |
|--------------- |---------:|--------:|---------:|----------:|
| Set200x200Formula | 433.1 ms | 8.35 ms | 13.00 ms | 443.21 MB |